+++ src/px-script/main/action-list/index.js
@@ -1,12 +1,10 @@
-import { isWithinBoundingBox } from "../utils";
-
 class ActionList {
     constructor (element) {
         this.container = element;
         this.toggleBtn = element.querySelector("i.material-icons");
         this.actionMenu = element.querySelector(".action-menu");
         this.actionMenuLinks = this.actionMenu.querySelectorAll("a");
-        this.isOpen = this.actionMenu.classList.contains("active");
+        this.isOpen = this.container.classList.contains("active");
 
         this.toggleBtn.addEventListener("click", () => {
             this.toggleMenu();
@@ -17,21 +15,17 @@ class ActionList {
     }
 
     open () {
-        this.actionMenu.classList.add("active");
+        this.container.classList.add("active");
         this.isOpen = true;
     }
 
     close () {
-        this.actionMenu.classList.remove("active");
+        this.container.classList.remove("active");
         this.isOpen = false;
     }
 
     toggleMenu () {
         this.isOpen ? this.close() : this.open();
-    }
-
-    containsPoint (x, y) {
-        return (isWithinBoundingBox(x, y, this.toggleBtn) || isWithinBoundingBox(x, y, this.actionMenu));
     }
 }
 
@@ -44,7 +38,7 @@ const actionList = (() => {
 
             document.addEventListener("click", e => {
                 actionLists.forEach(l => {
-                    if (!l.containsPoint(e.clientX, e.clientY) && l.isOpen) {
+                    if (e.target.closest(".action-list") !== l.container && l.isOpen) {
                         l.close();
                     }
                 });